+++ src/main/java/roomescape/controller/AdminController.java
@@ -11,8 +11,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import roomescape.controller.request.AdminReservationRequest;
-import roomescape.controller.response.ReservationResponse;
+import roomescape.dto.request.AdminReservationRequest;
+import roomescape.dto.response.ReservationResponse;
 import roomescape.service.ReservationService;
 
 @Controller
