+++ src/main/java/roomescape/controller/ReservationController.java
@@ -11,8 +11,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.controller.request.UserReservationRequest;
-import roomescape.controller.response.ReservationResponse;
+import roomescape.dto.request.UserReservationRequest;
+import roomescape.dto.response.ReservationResponse;
 import roomescape.service.ReservationService;
 import roomescape.ui.MemberIdConverter;
