+++ src/main/java/roomescape/dto/response/CheckMemberResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.response;
+package roomescape.dto.response;
 
 public record CheckMemberResponse(String name) {
 }
