+++ src/main/java/roomescape/dto/response/MemberResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.response;
+package roomescape.dto.response;
 
 import roomescape.domain.Member;
