+++ src/main/java/roomescape/service/ReservationService.java
@@ -4,9 +4,9 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.controller.request.AdminReservationRequest;
-import roomescape.controller.request.UserReservationRequest;
-import roomescape.controller.response.ReservationResponse;
+import roomescape.dto.request.AdminReservationRequest;
+import roomescape.dto.request.UserReservationRequest;
+import roomescape.dto.response.ReservationResponse;
 import roomescape.domain.Member;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
