+++ src/test/java/roomescape/TestFixtures.java
@@ -3,12 +3,7 @@ package roomescape;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
-import roomescape.controller.request.ReservationTimeRequest;
-import roomescape.controller.request.ThemeRequest;
-import roomescape.controller.request.UserReservationRequest;
-import roomescape.controller.response.ReservationResponse;
-import roomescape.controller.response.ReservationTimeResponse;
-import roomescape.controller.response.ThemeResponse;
+import roomescape.dto.response.ThemeResponse;
 import roomescape.domain.Member;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
