+++ src/test/java/roomescape/acceptance/ReservationTimeAcceptanceTest.java
@@ -17,7 +17,7 @@ import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.TestFactory;
 import org.springframework.test.context.jdbc.Sql;
 import roomescape.BasicAcceptanceTest;
-import roomescape.controller.response.AvailableReservationTimeResponse;
+import roomescape.dto.response.AvailableReservationTimeResponse;
 
 class ReservationTimeAcceptanceTest extends BasicAcceptanceTest {
     @TestFactory
